added AddItemForm.tsx, refactoring code
